fileupload/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pika==1.2.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
yarl==1.7.2
mysqlclient
django-mysql
